LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-canvas-tainting.https-expected.txt

CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
<<<<<<< HEAD
=======
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
>>>>>>> 19545d5ffb3... fix-179404
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: line 39: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: line 39: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
CONSOLE MESSAGE: line 39: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
CONSOLE MESSAGE: line 39: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
CONSOLE MESSAGE: line 39: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
<<<<<<< HEAD
CONSOLE MESSAGE: line 39: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
=======
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
>>>>>>> 19545d5ffb3... fix-179404
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
 

FAIL Verify canvas tainting of fetched image in a Service Worker assert_equals: expected "finish" but got "failure:Result of url:https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE&mode=same-origin&url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FPNGIMAGE  cross_origin:  must be NOT_TAINTED but TAINTED"
